xUnit fixtures are initialized even when the tests are all skipped, so changing tact for now.
