+++ tests/Aspire.Hosting.Testing.Tests/TestingBuilderTests.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Net.Http.Json;
 using Aspire.Hosting.Tests.Helpers;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
@@ -41,5 +42,23 @@ public class TestingBuilderTests
         var appModel = app.Services.GetRequiredService<DistributedApplicationModel>();
         Assert.Contains(appModel.GetContainerResources(), c => c.Name == "redis1");
         Assert.Contains(appModel.GetProjectResources(), p => p.Name == "myworker1");
+    }
+
+    [LocalOnlyFact]
+    public async Task HttpClientGetTest()
+    {
+        var builder = new DistributedApplicationTestingBuilder<Program>();
+        await using var app = builder.Build();
+        await app.StartAsync();
+
+        var httpClient = app.CreateHttpClient("mywebapp1");
+        var result1 = await httpClient.GetFromJsonAsync<WeatherForecast[]>("/weatherforecast");
+        Assert.NotNull(result1);
+        Assert.True(result1.Length > 0);
+    }
+
+    private sealed record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
+    {
+        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
     }
 }
